Moved firebase check to firebase module
