+++ src/auto_flutter_dig/task/setup/setup.py
@@ -4,7 +4,7 @@ from ...core.config import Config
 from ...model.argument import Args
 from ...model.task import *
 from ...task.identity import AflutterTaskIdentity
-from ..firebase.check import FirebaseCheck
+from ...module.firebase.task.setup.check import FirebaseCheck
 from ..flutter import FlutterCheck
 from .edit import SetupEdit
